jeff_agent/pyproject.toml
[project]
name = "jeff-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.25",
    "google-adk>=1.27.2",
    "groq>=0.37.1",
    "httpx>=0.28.1",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
